refactor: extract steer demo from AgentWorkspace
